--- .babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["@babel/preset-env"],
-    "plugins": [["@babel/transform-runtime"], "@babel/plugin-transform-strict-mode"]
-}
