--- .sequelizerc
@@ -1,8 +1,0 @@
-const path = require("path");
-
-module.exports = {
-    "config": path.resolve('./src/server/config/config.js'),
-    "models-path": path.resolve('./src/server/models'),
-    "seeders-path": path.resolve('./src/server/seeders'),
-    "migrations-path": path.resolve('./src/server/migrations') 
-};
