+++ README.md
@@ -1,26 +1,1 @@
- ## Guest Management System
-Guest Management System Api is built with express.js and postgresql. It allow guests to save their information while super admin retrieve, update and delete guest record. 
-
-It also allow user to view guest information.
-
-# Dependencies
-- express
-- express-async-handler
-- pg
-- cors
-- dotenv
-- sequelize
-- nodemon (dev)
-
-# Features 
-- Creating Guest 
-- Getting List of Guest
-- Updating Guest
-- Retrieving Guest information based on unique Id
-- Deleting Guest 
-
-# Guest Information
-- name
-- email 
-- phone number
-- gender 
+# management
