+++ package.json
@@ -2,9 +2,9 @@
   "name": "management",
   "version": "1.0.0",
   "description": "Guest Management System",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "nodemon --exec babel-node ./src/index.js"
+    "start": "nodemon server.js"
   },
   "repository": "https://github.com/offiongbassey/management.git",
   "author": "Offiong Bassey",
@@ -19,23 +19,9 @@
     "express-async-handler": "^1.2.0",
     "jsonwebtoken": "^9.0.1",
     "pg": "^8.11.1",
-    "sequelize": "^6.32.1",
-    "sequelize-cli": "^6.6.1"
+    "sequelize": "^6.32.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.22.9",
-    "@babel/core": "^7.22.9",
-    "@babel/eslint-parser": "^7.22.9",
-    "@babel/node": "^7.22.6",
-    "@babel/plugin-transform-runtime": "^7.22.9",
-    "@babel/plugin-transform-strict-mode": "^7.22.5",
-    "@babel/preset-env": "^7.22.9",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^9.1.3",
-    "babel-plugin-transform-strict-mode": "^6.24.1",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015-node": "^6.1.1",
-    "eslint": "^8.45.0",
     "nodemon": "^3.0.1"
   }
 }
