+++ server/config/Database.js
@@ -1,0 +1,14 @@
+const path = require("path");
+require("dotenv").config({path: path.resolve(__dirname, '../../.env')});
+const Sequelize = require("sequelize");
+const sequelize = new Sequelize(
+    process.env.DB_NAME,
+    process.env.DB_USER,
+    process.env.DB_PASSWORD,
+    {
+        host: process.env.DB_HOST,
+        dialect: "postgres"
+    }
+);
+
+module.exports = sequelize;
